+++ ios/RNFFmpegModule.m
@@ -1,27 +1,23 @@
 /*
- * MIT License
+ * Copyright (c) 2018-2020 Taner Sener
  *
- * Copyright (c) 2018 Taner Sener
+ * This file is part of ReactNativeFFmpeg.
  *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
+ * ReactNativeFFmpeg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
+ * ReactNativeFFmpeg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
  *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with ReactNativeFFmpeg.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#import "RNExecuteDelegate.h"
 #import "RNFFmpegModule.h"
 #import <React/RCTLog.h>
 #import <React/RCTBridge.h>
@@ -33,17 +29,11 @@
 
 static NSString *const PLATFORM_NAME = @"ios";
 
-static NSString *const KEY_VERSION = @"version";
-static NSString *const KEY_RC = @"rc";
-static NSString *const KEY_PLATFORM = @"platform";
-static NSString *const KEY_PACKAGE_NAME = @"packageName";
-static NSString *const KEY_LAST_RC = @"lastRc";
-static NSString *const KEY_LAST_COMMAND_OUTPUT = @"lastCommandOutput";
-static NSString *const KEY_PIPE = @"pipe";
-
+static NSString *const KEY_LOG_EXECUTION_ID = @"executionId";
 static NSString *const KEY_LOG_TEXT = @"log";
 static NSString *const KEY_LOG_LEVEL = @"level";
 
+static NSString *const KEY_STAT_EXECUTION_ID = @"executionId";
 static NSString *const KEY_STAT_TIME = @"time";
 static NSString *const KEY_STAT_SIZE = @"size";
 static NSString *const KEY_STAT_BITRATE = @"bitrate";
@@ -52,8 +42,13 @@ static NSString *const KEY_STAT_VIDEO_FRAME_NUMBER = @"videoFrameNumber";
 static NSString *const KEY_STAT_VIDEO_QUALITY = @"videoQuality";
 static NSString *const KEY_STAT_VIDEO_FPS = @"videoFps";
 
+static NSString *const KEY_EXECUTION_ID = @"executionId";
+static NSString *const KEY_EXECUTION_START_TIME = @"startTime";
+static NSString *const KEY_EXECUTION_COMMAND = @"command";
+
 static NSString *const EVENT_LOG = @"RNFFmpegLogCallback";
 static NSString *const EVENT_STAT = @"RNFFmpegStatisticsCallback";
+static NSString *const EVENT_EXECUTE = @"RNFFmpegExecuteCallback";
 
 @implementation RNFFmpegModule
 
@@ -64,46 +59,47 @@ RCT_EXPORT_MODULE(RNFFmpegModule);
 
     [array addObject:EVENT_LOG];
     [array addObject:EVENT_STAT];
+    [array addObject:EVENT_EXECUTE];
 
     return array;
 }
 
 RCT_EXPORT_METHOD(getPlatform:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
     NSString *architecture = [ArchDetect getArch];
-    resolve([RNFFmpegModule toStringDictionary:KEY_PLATFORM :[NSString stringWithFormat:@"%@-%@", PLATFORM_NAME, architecture]]);
+    resolve([NSString stringWithFormat:@"%@-%@", PLATFORM_NAME, architecture]);
 }
 
 RCT_EXPORT_METHOD(getFFmpegVersion:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
     NSString *ffmpegVersion = [MobileFFmpegConfig getFFmpegVersion];
-    resolve([RNFFmpegModule toStringDictionary:KEY_VERSION :ffmpegVersion]);
+    resolve(ffmpegVersion);
 }
 
 RCT_EXPORT_METHOD(executeFFmpegWithArguments:(NSArray*)arguments resolver:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
-    RCTLogInfo(@"Running FFmpeg with arguments: %@.\n", arguments);
-
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
         int rc = [MobileFFmpeg executeWithArguments:arguments];
-
-        RCTLogInfo(@"FFmpeg exited with rc: %d\n", rc);
-
-        resolve([RNFFmpegModule toIntDictionary:KEY_RC :[NSNumber numberWithInt:rc]]);
+        resolve([NSNumber numberWithInt:rc]);
     });
 }
 
-RCT_EXPORT_METHOD(executeFFprobeWithArguments:(NSArray*)arguments resolver:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
-    RCTLogInfo(@"Running FFprobe with arguments: %@.\n", arguments);
+RCT_EXPORT_METHOD(executeFFmpegAsyncWithArguments:(NSArray*)arguments resolver:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
+    RNExecuteDelegate* executeDelegate = [[RNExecuteDelegate alloc] initWithEventEmitter:self];
+    long executionId = [MobileFFmpeg executeWithArgumentsAsync:arguments withCallback:executeDelegate];
+    resolve([NSNumber numberWithLong:executionId]);
+}
 
+RCT_EXPORT_METHOD(executeFFprobeWithArguments:(NSArray*)arguments resolver:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
         int rc = [MobileFFprobe executeWithArguments:arguments];
-
-        RCTLogInfo(@"FFprobe exited with rc: %d\n", rc);
-
-        resolve([RNFFmpegModule toIntDictionary:KEY_RC :[NSNumber numberWithInt:rc]]);
+        resolve([NSNumber numberWithInt:rc]);
     });
 }
 
 RCT_EXPORT_METHOD(cancel) {
     [MobileFFmpeg cancel];
+}
+
+RCT_EXPORT_METHOD(cancelExecution:(int)executionId) {
+    [MobileFFmpeg cancel:executionId];
 }
 
 RCT_EXPORT_METHOD(enableRedirection) {
@@ -116,7 +112,7 @@ RCT_EXPORT_METHOD(disableRedirection) {
 
 RCT_EXPORT_METHOD(getLogLevel:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
     int logLevel = [MobileFFmpegConfig getLogLevel];
-    resolve([RNFFmpegModule toIntDictionary:KEY_LOG_LEVEL :[NSNumber numberWithInt:logLevel]]);
+    resolve([NSNumber numberWithInt:logLevel]);
 }
 
 RCT_EXPORT_METHOD(setLogLevel:(int)logLevel) {
@@ -158,7 +154,7 @@ RCT_EXPORT_METHOD(setFontDirectory:(NSString*)fontDirectoryPath with:(NSDictiona
 
 RCT_EXPORT_METHOD(getPackageName:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
     NSString *packageName = [MobileFFmpegConfig getPackageName];
-    resolve([RNFFmpegModule toStringDictionary:KEY_PACKAGE_NAME :packageName]);
+    resolve(packageName);
 }
 
 RCT_EXPORT_METHOD(getExternalLibraries:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
@@ -168,17 +164,15 @@ RCT_EXPORT_METHOD(getExternalLibraries:(RCTPromiseResolveBlock)resolve rejecter:
 
 RCT_EXPORT_METHOD(getLastReturnCode:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
     int lastReturnCode = [MobileFFmpegConfig getLastReturnCode];
-    resolve([RNFFmpegModule toIntDictionary:KEY_LAST_RC :[NSNumber numberWithInt:lastReturnCode]]);
+    resolve([NSNumber numberWithInt:lastReturnCode]);
 }
 
 RCT_EXPORT_METHOD(getLastCommandOutput:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
     NSString *lastCommandOutput = [MobileFFmpegConfig getLastCommandOutput];
-    resolve([RNFFmpegModule toStringDictionary:KEY_LAST_COMMAND_OUTPUT :lastCommandOutput]);
+    resolve(lastCommandOutput);
 }
 
 RCT_EXPORT_METHOD(getMediaInformation:(NSString*)path resolver:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
-    RCTLogInfo(@"Getting media information for %@.\n", path);
-
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
         MediaInformation *mediaInformation = [MobileFFprobe getMediaInformation:path];
         resolve([RNFFmpegModule toMediaInformationDictionary:mediaInformation]);
@@ -187,12 +181,22 @@ RCT_EXPORT_METHOD(getMediaInformation:(NSString*)path resolver:(RCTPromiseResolv
 
 RCT_EXPORT_METHOD(registerNewFFmpegPipe:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
     NSString *pipe = [MobileFFmpegConfig registerNewFFmpegPipe];
-    resolve([RNFFmpegModule toStringDictionary:KEY_PIPE :pipe]);
+    resolve(pipe);
 }
 
-- (void)logCallback: (int)level :(NSString*)message {
+RCT_EXPORT_METHOD(setEnvironmentVariable:(NSString*)variableName with:(NSString*)variableValue) {
+    setenv([variableName UTF8String], [variableValue UTF8String], true);
+}
+
+RCT_EXPORT_METHOD(listExecutions:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject) {
+    NSArray* executionsArray = [RNFFmpegModule toExecutionsArray:[MobileFFmpeg listExecutions]];
+    resolve(executionsArray);
+}
+
+- (void)logCallback:(long)executionId :(int)level :(NSString*)message {
     dispatch_async(dispatch_get_main_queue(), ^{
         NSMutableDictionary *dictionary = [[NSMutableDictionary alloc] init];
+        dictionary[KEY_LOG_EXECUTION_ID] = [NSNumber numberWithLong:executionId];
         dictionary[KEY_LOG_LEVEL] = [NSNumber numberWithInt:level];
         dictionary[KEY_LOG_TEXT] = message;
 
@@ -215,24 +219,12 @@ RCT_EXPORT_METHOD(registerNewFFmpegPipe:(RCTPromiseResolveBlock)resolve rejecter
     [self sendEventWithName:EVENT_STAT body:dictionary];
 }
 
-+ (NSDictionary *)toStringDictionary:(NSString*)key :(NSString*)value {
-    NSMutableDictionary *dictionary = [[NSMutableDictionary alloc] init];
-    dictionary[key] = value;
-
-    return dictionary;
-}
-
-+ (NSDictionary *)toIntDictionary:(NSString*)key :(NSNumber*)value {
-    NSMutableDictionary *dictionary = [[NSMutableDictionary alloc] init];
-    dictionary[key] = value;
-
-    return dictionary;
-}
-
 + (NSDictionary *)toStatisticsDictionary:(Statistics*)statistics {
     NSMutableDictionary *dictionary = [[NSMutableDictionary alloc] init];
 
     if (statistics != nil) {
+        dictionary[KEY_STAT_EXECUTION_ID] = [NSNumber numberWithLong: [statistics getExecutionId]];
+
         dictionary[KEY_STAT_TIME] = [NSNumber numberWithInt: [statistics getTime]];
         dictionary[KEY_STAT_SIZE] = [NSNumber numberWithLong: [statistics getSize]];
 
@@ -247,117 +239,32 @@ RCT_EXPORT_METHOD(registerNewFFmpegPipe:(RCTPromiseResolveBlock)resolve rejecter
     return dictionary;
 }
 
++ (NSArray *)toExecutionsArray:(NSArray*)ffmpegExecutions {
+    NSMutableArray *executions = [[NSMutableArray alloc] init];
+
+    for (int i = 0; i < [ffmpegExecutions count]; i++) {
+        FFmpegExecution* execution = [ffmpegExecutions objectAtIndex:i];
+
+        NSMutableDictionary *executionDictionary = [[NSMutableDictionary alloc] init];
+        executionDictionary[KEY_EXECUTION_ID] = [NSNumber numberWithLong: [execution getExecutionId]];
+        executionDictionary[KEY_EXECUTION_START_TIME] = [NSNumber numberWithDouble:[[execution getStartTime] timeIntervalSince1970]*1000];
+        executionDictionary[KEY_EXECUTION_COMMAND] = [execution getCommand];
+
+        [executions addObject: executionDictionary];
+    }
+
+    return executions;
+}
+
 + (NSDictionary *)toMediaInformationDictionary:(MediaInformation*)mediaInformation {
     NSMutableDictionary *dictionary = [[NSMutableDictionary alloc] init];
 
     if (mediaInformation != nil) {
-        if ([mediaInformation getFormat] != nil) {
-            dictionary[@"format"] =  [mediaInformation getFormat];
-        }
-        if ([mediaInformation getPath] != nil) {
-            dictionary[@"path"] = [mediaInformation getPath];
-        }
-        if ([mediaInformation getStartTime] != nil) {
-            dictionary[@"startTime"] = [mediaInformation getStartTime];
-        }
-        if ([mediaInformation getDuration] != nil) {
-            dictionary[@"duration"] = [mediaInformation getDuration];
-        }
-        if ([mediaInformation getBitrate] != nil) {
-            dictionary[@"bitrate"] = [mediaInformation getBitrate];
-        }
-        if ([mediaInformation getRawInformation] != nil) {
-            dictionary[@"rawInformation"] = [mediaInformation getRawInformation];
-        }
-
-        NSDictionary *metadataDictionary = [mediaInformation getMetadataEntries];
-        if (metadataDictionary != nil && ([metadataDictionary count] > 0)) {
-            dictionary[@"metadata"] = metadataDictionary;
-        }
-
-        NSArray *streams = [mediaInformation getStreams];
-        if (streams != nil && ([streams count] > 0)) {
-            NSMutableArray *array = [[NSMutableArray alloc] init];
-
-            for (int i=0; i < [streams count]; i++) {
-                StreamInformation *streamInformation= [streams objectAtIndex:i];
-                [array addObject: [RNFFmpegModule toStreamInformationDictionary:streamInformation]];
+        NSDictionary* allProperties = [mediaInformation getAllProperties];
+        if (allProperties != nil) {
+            for(NSString *key in [allProperties allKeys]) {
+                dictionary[key] = [allProperties objectForKey:key];
             }
-
-            dictionary[@"streams"] = array;
-        }
-    }
-
-    return dictionary;
-}
-
-+ (NSDictionary *)toStreamInformationDictionary:(StreamInformation*)streamInformation {
-    NSMutableDictionary *dictionary = [[NSMutableDictionary alloc] init];
-
-    if (streamInformation != nil) {
-        if ([streamInformation getIndex] != nil) {
-            dictionary[@"index"] = [streamInformation getIndex];
-        }
-        if ([streamInformation getType] != nil) {
-            dictionary[@"type"] = [streamInformation getType];
-        }
-        if ([streamInformation getCodec] != nil) {
-            dictionary[@"codec"] = [streamInformation getCodec];
-        }
-        if ([streamInformation getFullCodec] != nil) {
-            dictionary[@"fullCodec"] = [streamInformation getFullCodec];
-        }
-        if ([streamInformation getFormat] != nil) {
-            dictionary[@"format"] = [streamInformation getFormat];
-        }
-        if ([streamInformation getFullFormat] != nil) {
-            dictionary[@"fullFormat"] = [streamInformation getFullFormat];
-        }
-        if ([streamInformation getWidth] != nil) {
-            dictionary[@"width"] = [streamInformation getWidth];
-        }
-        if ([streamInformation getHeight] != nil) {
-            dictionary[@"height"] = [streamInformation getHeight];
-        }
-        if ([streamInformation getBitrate] != nil) {
-            dictionary[@"bitrate"] = [streamInformation getBitrate];
-        }
-        if ([streamInformation getSampleRate] != nil) {
-            dictionary[@"sampleRate"] = [streamInformation getSampleRate];
-        }
-        if ([streamInformation getSampleFormat] != nil) {
-            dictionary[@"sampleFormat"] = [streamInformation getSampleFormat];
-        }
-        if ([streamInformation getChannelLayout] != nil) {
-            dictionary[@"channelLayout"] = [streamInformation getChannelLayout];
-        }
-        if ([streamInformation getSampleAspectRatio] != nil) {
-            dictionary[@"sampleAspectRatio"] = [streamInformation getSampleAspectRatio];
-        }
-        if ([streamInformation getDisplayAspectRatio] != nil) {
-            dictionary[@"displayAspectRatio"] = [streamInformation getDisplayAspectRatio];
-        }
-        if ([streamInformation getAverageFrameRate] != nil) {
-            dictionary[@"averageFrameRate"] = [streamInformation getAverageFrameRate];
-        }
-        if ([streamInformation getRealFrameRate] != nil) {
-            dictionary[@"realFrameRate"] = [streamInformation getRealFrameRate];
-        }
-        if ([streamInformation getTimeBase] != nil) {
-            dictionary[@"timeBase"] = [streamInformation getTimeBase];
-        }
-        if ([streamInformation getCodecTimeBase] != nil) {
-            dictionary[@"codecTimeBase"] = [streamInformation getCodecTimeBase];
-        }
-
-        NSDictionary *metadataDictionary = [streamInformation getMetadataEntries];
-        if (metadataDictionary != nil && ([metadataDictionary count] > 0)) {
-            dictionary[@"metadata"] = metadataDictionary;
-        }
-
-        NSDictionary *sidedataDictionary = [streamInformation getSidedataEntries];
-        if (sidedataDictionary != nil && ([sidedataDictionary count] > 0)) {
-            dictionary[@"sidedata"] = sidedataDictionary;
         }
     }